+++ Source/Terrain/SplineItems.c
@@ -28,7 +28,7 @@ static Boolean NilPrime(long splineNum, SplineItemType *itemPtr);
 /**********************/
 
 SplineDefType	**gSplineList = nil;
-long			gNumSplines = 0;
+int				gNumSplines = 0;
 
 static long		gNumSplineObjects = 0;
 static ObjNode	*gSplineObjectList[MAX_SPLINE_OBJECTS];
